GVT copyrigh,utilt changes
