+++ dev/com.ibm.ws.gvt.rest_fat/publish/servers/com.ibm.gvt.server/server.xml
@@ -1,5 +1,5 @@
 <!--
-    Copyright (c) 2021, 2024 IBM Corporation and others.
+    Copyright (c) 2024 IBM Corporation and others.
     All rights reserved. This program and the accompanying materials
     are made available under the terms of the Eclipse Public License 2.0
     which accompanies this distribution, and is available at
