+++ dev/fattest.simplicity/bnd.bnd
@@ -117,4 +117,4 @@ fat.test.container.images: testcontainers/ryuk:0.6.0, testcontainers/sshd:1.1.0,
     org.bouncycastle:bcpkix-jdk18on;version=1.75,\
     org.bouncycastle:bcprov-jdk18on;version=1.75,\
     com.ibm.ws.org.apache.httpcomponents;version=latest,\
-    org.apache.httpcomponents:httpcore;version=latest
+    org.apache.httpcomponents:httpcore;version=4.1.2
